requirements.txt
openpyxl==3.0.10
pymongo==3.6.0
pendulum==2.1.2
xmltodict==0.13.0
pandas==1.3.5
requests==2.28.1
apache-airflow-providers-common-sql==1.3.0
apache-airflow-providers-mongo==3.1.0